update clear_db script

+++ clear_db.py
@@ -2,10 +2,19 @@ import sqlite3 as sq
 
 conn = sq.connect("rfid_access.db")
 curs = conn.cursor()
-curs.execute("DROP TABLE users")
-curs.execute("DROP TABLE daily")
-curs.execute("DROP TABLE annual")
-curs.execute("DROP TABLE log")
+ans = input("Delete users table? (y)").lower()
+if ans == "y":
+    try:
+        curs.execute("DROP TABLE users")
+    except sq.OperationalError:
+        pass
+
+table = ["daily", "annual", "log"]
+for t in table:
+    try:
+        curs.execute("DROP TABLE " + t)
+    except sq.OperationalError:
+        pass
 conn.commit()
 
 conn.close()
